the first assignment was to write a program to print the first 100 Fibonacci numbers

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,7 +9,7 @@
 
 int main(void)
 {
-	unsigned long long int num1, num2, n, num1A, num2A, num1B, num2B, count;
+	unsigned long int num1, num2, n, num1A, num2A, num1B, num2B, count;
 
 	num1 = 1;
 	num2 = 2;
